feat: implemented cookies and asyncronous loading of the page.
now it ist possible to refresh the page without loosing the state of the game.
When you load the page, first of all you load the empty page without any gamestate.
if you dont have cookies, you send a second request for a gameID and a playerID by a GET request.
The response of the server will contain a new GameID and PlayerID
if you already have cookies, you send youre cookies to the syrver by a POST request.
The response of the server will contain the state of your game in form of a FEN string (Forsyth-Edwards-Notation)